* 'feature/Q&A' of https://github.com/zhiyicx/thinksns-plus-android:
  (#669) 修改话题专家列表的样式 展示话题详情数据
  (#669) 添加获取话题的问答列表、修改话题关注状态的接口
  (#669) 添加话题详情的接口
  (#669) 添加话题详情页
  (#669) 话题详情 添加单独的头像列表

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/q_a/publish/add_topic/AddTopicFragment.java
@@ -79,7 +79,7 @@ public class AddTopicFragment extends TSListFragment<AddTopicContract.Presenter,
         List<QAPublishBean.Topic> topicList = new ArrayList<>();
         for (QATopicBean qaTopicBean : mQATopicBeanList) {
             QAPublishBean.Topic topic = new QAPublishBean.Topic();
-            topic.setId(qaTopicBean.getId());
+            topic.setId(qaTopicBean.getId().intValue());
             topicList.add(topic);
         }
         mQAPublishBean.setTopics(topicList);
